OT249-Tests: added new tests
