+++ src/main/java/com/alkemy/ong/controller/BucketController.java
@@ -1,7 +1,7 @@
 
 package com.alkemy.ong.controller;
 
-import com.alkemy.ong.service.AmazonClient;
+import com.alkemy.ong.service.impl.AmazonClientImpl;
 import lombok.AllArgsConstructor;
 
 import org.springframework.web.bind.annotation.PostMapping;
@@ -13,10 +13,10 @@ import org.springframework.web.multipart.MultipartFile;
 @AllArgsConstructor
 public class BucketController {
 
-    private final AmazonClient amazonClient;
+    private final AmazonClientImpl amazonClientImpl;
 
     @PostMapping("/uploadFile")
     public String uploadFile(@RequestPart(value = "file") MultipartFile file) {
-        return amazonClient.uploadFile(file);
+        return amazonClientImpl.uploadFile(file);
     }
 }
