+++ src/main/java/com/alkemy/ong/service/impl/ActivityServiceImpl.java
@@ -2,6 +2,7 @@ package com.alkemy.ong.service.impl;
 
 import java.util.Optional;
 
+import lombok.AllArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -10,6 +11,7 @@ import com.alkemy.ong.repository.ActivityRepository;
 import com.alkemy.ong.service.ActivityService;
 
 @Service
+@AllArgsConstructor
 public class ActivityServiceImpl implements ActivityService {
 
 	@Autowired
