+++ src/main/java/com/alkemy/ong/service/impl/AmazonClientImpl.java
@@ -1,0 +1,88 @@
+package com.alkemy.ong.service.impl;
+
+import com.alkemy.ong.service.AmazonClient;
+import com.amazonaws.auth.AWSCredentials;
+import com.amazonaws.auth.AWSStaticCredentialsProvider;
+import com.amazonaws.auth.BasicAWSCredentials;
+import com.amazonaws.services.s3.AmazonS3;
+import com.amazonaws.services.s3.AmazonS3ClientBuilder;
+import com.amazonaws.services.s3.model.CannedAccessControlList;
+import com.amazonaws.services.s3.model.PutObjectRequest;
+
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Service;
+import org.springframework.web.multipart.MultipartFile;
+
+import javax.annotation.PostConstruct;
+import java.io.File;
+import java.io.FileOutputStream;
+import java.io.IOException;
+import java.util.Date;
+import java.util.Objects;
+
+@Service
+public class AmazonClientImpl implements AmazonClient {
+
+    private AmazonS3 amazonS3;
+
+    private final String endpointUrl;
+    private final String bucketName;
+    private final String accessKey;
+    private final String secretKey;
+
+    public AmazonClientImpl(@Value("${amazon.endpointUrl}") String endpointUrl,
+                            @Value("${amazon.bucketName}") String bucketName,
+                            @Value("${amazon.accessKey}") String accessKey,
+                            @Value("${amazon.secretKey}") String secretKey) {
+        this.endpointUrl = endpointUrl;
+        this.bucketName = bucketName;
+        this.accessKey = accessKey;
+        this.secretKey = secretKey;
+    }
+
+    @PostConstruct
+    private void initializeAmazon() {
+        AWSCredentials credentials = new BasicAWSCredentials(this.accessKey, this.secretKey);
+        this.amazonS3 = AmazonS3ClientBuilder
+                .standard()
+                .withRegion("us-east-1")
+                .withCredentials(new AWSStaticCredentialsProvider(credentials))
+                .build();
+    }
+
+    private File convertMultiPartToFIle(MultipartFile file) throws IOException {
+        File convertFile = new File(Objects.requireNonNull(file.getOriginalFilename()));
+        FileOutputStream fos = new FileOutputStream(convertFile);
+        fos.write(file.getBytes());
+        fos.close();
+        return convertFile;
+    }
+
+    private String generateFIleNAme(MultipartFile multipart) {
+        return new Date()
+                .getTime() + "-" +
+                multipart
+                        .getOriginalFilename()
+                        .replace(" ", "_");
+    }
+
+    private void uploadFileToS3Bucket(String fileName, File file) {
+        amazonS3.putObject(new PutObjectRequest(bucketName, fileName, file)
+                .withCannedAcl(CannedAccessControlList.PublicRead));
+    }
+
+    public String uploadFile(MultipartFile multipartFile) {
+        String fileUrl = "";
+        try {
+            File file = convertMultiPartToFIle(multipartFile);
+            String fileName = generateFIleNAme(multipartFile);
+            fileUrl = endpointUrl + "/" + bucketName + "/" + fileName;
+            uploadFileToS3Bucket(fileName, file);
+            file.delete();
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+        return fileUrl;
+
+    }
+}
